requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
python-dotenv==1.0.0
twilio==8.12.0
pyotp==2.9.0
email-validator==1.1.3
pymongo==4.6.1
flask-pymongo==2.3.0 